forecase_app/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
cmdstanpy==1.0.7
convertdate==2.4.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
ephem==4.1.3
fonttools==4.37.1
hijri-converter==2.2.4
holidays==0.15
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.3
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
prophet==1.1.1
psycopg2-binary==2.9.3
PyMeeus==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools-git==1.2
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.1
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.12
